requirements.txt
fastapi
uvicorn
torch
transformers
sentencepiece
pymupdf
faiss-cpu
numpy
requests
python-multipart